Update database.py
Because of the OR t.CreatedByAgentID = ? part, tickets that the agent created appear under “Meine Tickets” even if they are not assigned to them. To restrict the list to only tickets assigned to the agent, this condition should be changed to only check ta.AgentID = ? (removing the OR t.CreatedByAgentID = ? clause).